rename server side files to appointments, fix client side errors
