+++ client/app/admin-dashboard/page.jsx
@@ -1,6 +1,6 @@
 "use client";
 import { useState, useEffect } from "react";
-import ReservationTable from "../../components/ReservationTable";
+import ReservationTable from "../../components/admin_dashboard/ReservationTable";
 
 export default function AdminDashboardView() {
   const [appointments, setAppointments] = useState([]);
